requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
bcrypt==4.1.3; python_version >= '3.7'
certifi==2024.2.2; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
dnspython==2.6.1; python_version >= '3.8'
ecdsa==0.19.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
email-validator==2.1.1; python_version >= '3.8'
et-xmlfile==1.1.0; python_version >= '3.6'
fastapi==0.111.0; python_version >= '3.8'
fastapi-cli==0.0.4; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httptools==0.6.1
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
jinja2==3.1.4; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
numpy==1.26.4; python_version >= '3.12'
openpyxl==3.1.3; python_version >= '3.6'
orjson==3.10.3; python_version >= '3.8'
pandas==2.2.2; python_version >= '3.9'
passlib==1.7.4
pyasn1==0.6.0; python_version >= '3.8'
pydantic==2.7.2; python_version >= '3.8'
pydantic-core==2.18.3; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pymongo==4.7.2; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9; python_version >= '3.8'
pytz==2024.1
pyyaml==6.0.1
rich==13.7.1; python_full_version >= '3.7.0'
rsa==4.9; python_version >= '3.6' and python_version < '4'
shellingham==1.5.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
typer==0.12.3; python_version >= '3.7'
typing-extensions==4.12.0; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
ujson==5.10.0; python_version >= '3.8'
uvicorn[standard]==0.30.0; python_version >= '3.8'
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
